The main script has been modified into several modules
