+++ js/createElement.js
@@ -1,0 +1,8 @@
+export function createElement(tag, className) {
+  tag = document.createElement(tag);
+  if (className) {
+    tag.classList.add(className);
+  }
+  
+  return tag;
+};
