+++ js/getRandom.js
@@ -1,0 +1,3 @@
+export function getRandom(n) {
+  return Math.ceil(Math.random() * n);
+}
