+++ js/players.js
@@ -1,0 +1,37 @@
+import { changeHP, elHP, renderHP } from './playerHelpers.js'
+
+export const HIT = {
+    head: 30,
+    body: 25,
+    foot: 20,
+};
+
+export const ATTACK = ['head', 'body', 'foot'];
+
+export const player1 = {
+  player: 1,
+  changeHP,
+  renderHP,
+  elHP,
+  name: 'Scorpion',
+  hp: 100,
+  img: 'http://reactmarathon-api.herokuapp.com/assets/scorpion.gif',
+  weapon: ['sabre of light', 'sawblade', 'sword', 'sai', 'sektors rocket'],
+  attack: () => {
+    console.log(this.name + 'Fight...');
+  }
+};
+
+export const player2 = {
+  player: 2,
+  changeHP,
+  renderHP,
+  elHP,
+  name: 'Kitana',
+  hp: 100,
+  img: 'http://reactmarathon-api.herokuapp.com/assets/kitana.gif',
+  weapon: ['kitana', 'sword', 'sai', 'sento', 'saber teeth'],
+  attack: () => {
+    console.log(this.name + 'Fight...');
+  }
+};
