+++ js/showResult.js
@@ -1,0 +1,35 @@
+import { player1, player2 } from './players.js';
+import { createReloadButton } from './createReloadButton.js';
+import { createElement } from './createElement.js';
+import { logsCase } from './generateLogs.js';
+
+const arenas = document.querySelector('.arenas');
+
+function playerWins(name) {
+  const winsTitle = createElement('div', 'winsTitle');
+  if (name) {
+    winsTitle.innerText = `${name} wins`;
+  } else {
+    winsTitle.innerText = 'draw';
+  }
+  
+  return winsTitle;
+};
+
+export const showResult = () => {
+  if (player1.hp === 0 || player2.hp === 0) {
+    arenas.append(createReloadButton());
+  }
+
+  if (player1.hp === 0 && player1.hp < player2.hp) {
+    arenas.append(playerWins(player2.name));
+    logsCase('end', player2, player1);
+  } else if (player2.hp === 0 && player2.hp < player1.hp) {
+    arenas.append(playerWins(player1.name));
+    logsCase('end', player1, player2);
+
+  } else if (player1.hp === 0 && player2.hp === 0) {
+    arenas.append(playerWins());
+    logsCase('draw');
+  }
+};
